+++ tests/test_components/test_monitor.py
@@ -309,7 +309,7 @@ def test_mode_bend_radius():
             size=(5, 0, 1),
             freqs=np.linspace(1e14, 2e14, 100),
             name="test",
-            mode_spec=td.ModeSpec(num_modes=1, bend_radius=1, bend_axis=1),
+            mode_spec=td.ModeSpec(num_modes=1, bend_radius=1, bend_axis=2),
         )
         _ = td.Simulation(
             size=(2, 2, 2),