+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bug in contains check for `LumpedElement`, which should allow the case of a `LumpedElement` touching the simulation boundaries.
 - Bug when generating a grid with snapping points near the simulation boundaries.
 - Fixed field colocation in `EMEModeSolverMonitor`.
+- Solver error for EME simulations with bends, introduced when support for 2D EME simulations was added.
 
 ### Changed
 - Relaxed bounds checking of path integrals during `WavePort` validation.